fix: implement payment with appchain token
