+++ pallets/genetic-analysis-orders/src/interface.rs
@@ -9,13 +9,14 @@ pub trait GeneticAnalysisOrderInterface<T: frame_system::Config> {
 		price_index: u32,
 		customer_box_public_key: &T::Hash,
 		genetic_link: &[u8],
+		asset_id: Option<u32>,
 	) -> Result<Self::GeneticAnalysisOrder, Self::Error>;
 	fn cancel_genetic_analysis_order(
 		customer_id: &T::AccountId,
 		genetic_analysis_order_id: &T::Hash,
 	) -> Result<Self::GeneticAnalysisOrder, Self::Error>;
 	fn set_genetic_analysis_order_paid(
-		customer_id: &T::AccountId,
+		account_id: &T::AccountId,
 		genetic_analysis_order_id: &T::Hash,
 	) -> Result<Self::GeneticAnalysisOrder, Self::Error>;
 	fn fulfill_genetic_analysis_order(
