+++ pallets/genetic-analysis-orders/src/lib.rs
@@ -1,9 +1,13 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub use pallet::*;
+
+pub mod functions;
+pub mod impl_genetic_analysis_orders;
 pub mod interface;
 pub mod migrations;
+pub mod types;
 pub mod weights;
-use interface::GeneticAnalysisOrderInterface;
 
 use frame_support::{
 	codec::{Decode, Encode},
@@ -13,22 +17,17 @@ use frame_support::{
 		RuntimeDebug, SaturatedConversion,
 	},
 	sp_std::convert::TryInto,
-	traits::{fungibles, Currency, ExistenceRequirement, StorageVersion},
+	traits::{Currency, StorageVersion},
 	PalletId,
 };
-pub use pallet::*;
 use primitives_price_and_currency::{CurrencyType, Price};
 use primitives_tracking_id::TrackingId;
-pub use scale_info::TypeInfo;
 use sp_std::{prelude::*, vec};
 use traits_genetic_analysis::{GeneticAnalysisProvider, GeneticAnalysisTracking};
-use traits_genetic_analysis_orders::{
-	GeneticAnalysisOrderEventEmitter, GeneticAnalysisOrderStatusUpdater,
-};
+
 use traits_genetic_analyst_services::{GeneticAnalystServiceInfo, GeneticAnalystServicesProvider};
 use traits_genetic_analysts::GeneticAnalystsProvider;
 use traits_genetic_data::{GeneticData, GeneticDataProvider};
-pub use weights::WeightInfo;
 
 #[cfg(test)]
 mod mock;
@@ -36,111 +35,28 @@ mod mock;
 #[cfg(test)]
 mod tests;
 
-#[derive(Encode, Decode, Clone, RuntimeDebug, PartialEq, Eq, TypeInfo)]
-pub enum GeneticAnalysisOrderStatus {
-	Unpaid,
-	Paid,
-	Fulfilled,
-	Refunded,
-	Cancelled,
-	Failed,
-}
-impl Default for GeneticAnalysisOrderStatus {
-	fn default() -> Self {
-		GeneticAnalysisOrderStatus::Unpaid
-	}
-}
-
-#[derive(Encode, Decode, Clone, Default, RuntimeDebug, PartialEq, Eq, TypeInfo)]
-pub struct GeneticAnalysisOrder<Hash, AccountId, Balance, Moment> {
-	pub id: Hash,
-	pub service_id: Hash,
-	pub customer_id: AccountId,
-	pub customer_box_public_key: Hash,
-	pub seller_id: AccountId,
-	pub genetic_data_id: Hash,
-	pub genetic_analysis_tracking_id: TrackingId,
-	pub currency: CurrencyType,
-	pub prices: Vec<Price<Balance>>,
-	pub additional_prices: Vec<Price<Balance>>,
-	pub total_price: Balance,
-	pub status: GeneticAnalysisOrderStatus,
-	pub created_at: Moment,
-	pub updated_at: Moment,
-	pub genetic_link: Vec<u8>,
-}
-#[allow(clippy::too_many_arguments)]
-impl<Hash, AccountId, Balance, Moment: Default>
-	GeneticAnalysisOrder<Hash, AccountId, Balance, Moment>
-{
-	pub fn new(
-		id: Hash,
-		service_id: Hash,
-		customer_id: AccountId,
-		customer_box_public_key: Hash,
-		seller_id: AccountId,
-		genetic_data_id: Hash,
-		genetic_analysis_tracking_id: TrackingId,
-		genetic_link: Vec<u8>,
-		currency: CurrencyType,
-		prices: Vec<Price<Balance>>,
-		additional_prices: Vec<Price<Balance>>,
-		total_price: Balance,
-		created_at: Moment,
-	) -> Self {
-		Self {
-			id,
-			service_id,
-			customer_id,
-			customer_box_public_key,
-			seller_id,
-			genetic_data_id,
-			genetic_analysis_tracking_id,
-			genetic_link,
-			currency,
-			prices,
-			additional_prices,
-			status: GeneticAnalysisOrderStatus::default(),
-			total_price,
-			created_at,
-			updated_at: Moment::default(),
-		}
-	}
-
-	pub fn get_id(&self) -> &Hash {
-		&self.id
-	}
-
-	pub fn get_created_at(&self) -> &Moment {
-		&self.created_at
-	}
-
-	pub fn get_service_id(&self) -> &Hash {
-		&self.service_id
-	}
-}
+pub use interface::GeneticAnalysisOrderInterface;
+pub use types::*;
+pub use weights::WeightInfo;
 
 /// The current storage version.
-const STORAGE_VERSION: StorageVersion = StorageVersion::new(2);
-
-// Asset ID and Balance types
-pub type AssetId = u32;
-pub type AssetBalance = u128;
+const STORAGE_VERSION: StorageVersion = StorageVersion::new(3);
 
 #[frame_support::pallet]
 pub mod pallet {
-	use crate::*;
+	use super::*;
+
 	use frame_support::{dispatch::DispatchResultWithPostInfo, traits::tokens::fungibles};
 	use frame_system::pallet_prelude::*;
 
 	#[pallet::config]
 	pub trait Config: frame_system::Config + pallet_timestamp::Config {
 		type Event: From<Event<Self>> + IsType<<Self as frame_system::Config>::Event>;
-		type Assets: fungibles::Mutate<
-			<Self as frame_system::Config>::AccountId,
-			AssetId = AssetId,
-			Balance = AssetBalance,
-		>;
+		type Assets: fungibles::Transfer<
+				<Self as frame_system::Config>::AccountId,
+				AssetId = AssetId,
+				Balance = AssetBalance,
+			> + fungibles::InspectMetadata<<Self as frame_system::Config>::AccountId>;
 		type GeneticAnalysts: GeneticAnalystsProvider<Self>;
 		type GeneticAnalystServices: GeneticAnalystServicesProvider<Self, BalanceOf<Self>>;
 		type GeneticData: GeneticDataProvider<Self>;
@@ -327,6 +243,7 @@ pub mod pallet {
 		Token,
 		Arithmetic,
 		WrongAssetIdFormat,
+		AssetIdNotFound,
 	}
 
 	#[pallet::call]
@@ -339,6 +256,7 @@ pub mod pallet {
 			price_index: u32,
 			customer_box_public_key: T::Hash,
 			genetic_link: Vec<u8>,
+			asset_id: Option<u32>,
 		) -> DispatchResultWithPostInfo {
 			let who = ensure_signed(origin)?;
 
@@ -349,6 +267,7 @@ pub mod pallet {
 				price_index,
 				&customer_box_public_key,
 				&genetic_link,
+				asset_id,
 			) {
 				Ok(genetic_analysis_order) => {
 					Self::deposit_event(Event::<T>::GeneticAnalysisOrderCreated(
@@ -508,740 +427,5 @@ pub mod pallet {
 
 			Ok(Pays::No.into())
 		}
-	}
-}
-
-impl<T: Config> GeneticAnalysisOrderInterface<T> for Pallet<T> {
-	type GeneticAnalysisOrder = GeneticAnalysisOrderOf<T>;
-	type Error = Error<T>;
-
-	fn create_genetic_analysis_order(
-		customer_id: &T::AccountId,
-		genetic_data_id: &T::Hash,
-		genetic_analyst_service_id: &T::Hash,
-		price_index: u32,
-		customer_box_public_key: &T::Hash,
-		genetic_link: &[u8],
-	) -> Result<Self::GeneticAnalysisOrder, Self::Error> {
-		let genetic_analyst_service = match T::GeneticAnalystServices::genetic_analyst_service_by_id(
-			genetic_analyst_service_id,
-		) {
-			Some(_genetic_analyst_service) => _genetic_analyst_service,
-			None => return Err(Error::<T>::GeneticAnalystServiceDoesNotExist),
-		};
-
-		let seller_id = genetic_analyst_service.get_owner_id();
-		if !T::GeneticAnalysts::is_genetic_analyst_available(seller_id).unwrap() {
-			// If _bool is false, then genetic analyst is unavailable
-			return Err(Error::<T>::GeneticAnalystUnavailable)
-		}
-
-		let genetic_data = T::GeneticData::genetic_data_by_id(genetic_data_id);
-		if genetic_data.is_none() {
-			return Err(Error::<T>::GeneticDataDoesNotExist)
-		}
-
-		let genetic_data = genetic_data.unwrap();
-		if customer_id.clone() != genetic_data.get_owner_id().clone() {
-			return Err(Error::<T>::NotOwnerOfGeneticData)
-		}
-
-		let prices_by_currency = genetic_analyst_service.get_prices_by_currency();
-		if prices_by_currency.is_empty() ||
-			prices_by_currency.len() - 1 < price_index.try_into().unwrap()
-		{
-			return Err(Error::<T>::PriceIndexNotFound)
-		}
-
-		let price_by_currency = &prices_by_currency[price_index as usize];
-
-		let total_price = &price_by_currency.total_price;
-		let currency = &price_by_currency.currency;
-		let prices = &price_by_currency.price_components;
-		let additional_prices = &price_by_currency.additional_prices;
-
-		let now = pallet_timestamp::Pallet::<T>::get();
-
-		// Initialize GeneticAnalysis
-		let genetic_analysis_order_id =
-			Self::generate_genetic_analysis_order_id(customer_id, genetic_analyst_service_id);
-
-		let genetic_analysis = T::GeneticAnalysis::register_genetic_analysis(
-			seller_id,
-			customer_id,
-			&genetic_analysis_order_id,
-		);
-
-		if genetic_analysis.is_err() {
-			return Err(Error::<T>::GeneticAnalysisInitalizationError)
-		}
-		let genetic_analysis = genetic_analysis.ok().unwrap();
-
-		let genetic_analysis_order = GeneticAnalysisOrder::new(
-			genetic_analysis_order_id,
-			*genetic_analyst_service_id,
-			customer_id.clone(),
-			*customer_box_public_key,
-			seller_id.clone(),
-			*genetic_data_id,
-			genetic_analysis.get_genetic_analysis_tracking_id().clone(),
-			genetic_link.to_vec(),
-			currency.clone(),
-			prices.clone(),
-			additional_prices.clone(),
-			*total_price,
-			now,
-		);
-		Self::insert_genetic_analysis_order_to_storage(&genetic_analysis_order);
-
-		Ok(genetic_analysis_order)
-	}
-
-	fn cancel_genetic_analysis_order(
-		customer_id: &T::AccountId,
-		genetic_analysis_order_id: &T::Hash,
-	) -> Result<Self::GeneticAnalysisOrder, Self::Error> {
-		let genetic_analysis_order = GeneticAnalysisOrders::<T>::get(genetic_analysis_order_id);
-		if genetic_analysis_order.is_none() {
-			return Err(Error::<T>::GeneticAnalysisOrderNotFound)
-		}
-		let genetic_analysis_order = genetic_analysis_order.unwrap();
-
-		if genetic_analysis_order.customer_id != customer_id.clone() {
-			return Err(Error::<T>::UnauthorizedGeneticAnalysisOrderCancellation)
-		}
-
-		let genetic_analysis =
-			T::GeneticAnalysis::genetic_analysis_by_genetic_analysis_tracking_id(
-				&genetic_analysis_order.genetic_analysis_tracking_id,
-			)
-			.unwrap();
-		if !genetic_analysis.is_registered() {
-			return Err(Error::<T>::OngoingGeneticAnalysisOrderCannotBeCancelled)
-		}
-
-		// Default status would be cancelled
-		let mut genetic_analysis_order_status = GeneticAnalysisOrderStatus::Cancelled;
-
-		if genetic_analysis_order.status == GeneticAnalysisOrderStatus::Paid {
-			if !Self::is_pallet_balance_sufficient_for_transfer(genetic_analysis_order.total_price)
-			{
-				return Err(Error::<T>::InsufficientPalletFunds)
-			}
-
-			if genetic_analysis_order.currency == CurrencyType::DBIO {
-				match CurrencyOf::<T>::transfer(
-					&Self::account_id(),
-					&genetic_analysis_order.customer_id,
-					genetic_analysis_order.total_price,
-					ExistenceRequirement::AllowDeath,
-				) {
-					Ok(_) => (),
-					Err(dispatch) => match dispatch {
-						sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-						sp_runtime::DispatchError::CannotLookup =>
-							return Err(Error::<T>::CannotLookup),
-						sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-						sp_runtime::DispatchError::TooManyConsumers =>
-							return Err(Error::<T>::TooManyConsumers),
-						sp_runtime::DispatchError::ConsumerRemaining =>
-							return Err(Error::<T>::ConsumerRemaining),
-						sp_runtime::DispatchError::NoProviders =>
-							return Err(Error::<T>::NoProviders),
-						sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-						sp_runtime::DispatchError::Arithmetic(_) =>
-							return Err(Error::<T>::Arithmetic),
-						sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-					},
-				}
-			} else {
-				let asset_id = Self::asset_id(&genetic_analysis_order.currency)?;
-				match <T::Assets as fungibles::Mutate<T::AccountId>>::teleport(
-					asset_id,
-					&Self::account_id(),
-					&genetic_analysis_order.customer_id,
-					genetic_analysis_order.total_price.saturated_into(),
-				) {
-					Ok(_) => (),
-					Err(dispatch) => match dispatch {
-						sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-						sp_runtime::DispatchError::CannotLookup =>
-							return Err(Error::<T>::CannotLookup),
-						sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-						sp_runtime::DispatchError::TooManyConsumers =>
-							return Err(Error::<T>::TooManyConsumers),
-						sp_runtime::DispatchError::ConsumerRemaining =>
-							return Err(Error::<T>::ConsumerRemaining),
-						sp_runtime::DispatchError::NoProviders =>
-							return Err(Error::<T>::NoProviders),
-						sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-						sp_runtime::DispatchError::Arithmetic(_) =>
-							return Err(Error::<T>::Arithmetic),
-						sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-					},
-				}
-			}
-
-			// If code reaches here change status to Refunded
-			genetic_analysis_order_status = GeneticAnalysisOrderStatus::Refunded;
-		}
-
-		Self::remove_genetic_analysis_order_id_from_pending_genetic_analysis_orders_by_seller(
-			&genetic_analysis_order.seller_id,
-			genetic_analysis_order_id,
-		);
-
-		// Delete dna sample associated with the genetic_analysis_order
-		let _genetic_analysis = T::GeneticAnalysis::delete_genetic_analysis(
-			&genetic_analysis_order.genetic_analysis_tracking_id,
-		);
-
-		let genetic_analysis_order = Self::update_genetic_analysis_order_status(
-			genetic_analysis_order_id,
-			genetic_analysis_order_status,
-		)
-		.unwrap();
-
-		Ok(genetic_analysis_order)
-	}
-
-	fn set_genetic_analysis_order_paid(
-		customer_id: &T::AccountId,
-		genetic_analysis_order_id: &T::Hash,
-	) -> Result<Self::GeneticAnalysisOrder, Self::Error> {
-		let genetic_analysis_order = GeneticAnalysisOrders::<T>::get(genetic_analysis_order_id)
-			.ok_or(Error::<T>::GeneticAnalysisOrderNotFound)?;
-
-		if customer_id != &genetic_analysis_order.customer_id {
-			let _ = EscrowKey::<T>::get()
-				.filter(|account_id| account_id == customer_id)
-				.ok_or(Error::<T>::Unauthorized)?;
-		}
-
-		if !Self::is_balance_sufficient_for_payment(
-			&genetic_analysis_order.customer_id,
-			genetic_analysis_order.total_price,
-		) {
-			return Err(Error::<T>::InsufficientFunds)
-		}
-
-		if genetic_analysis_order.currency == CurrencyType::DBIO {
-			match CurrencyOf::<T>::transfer(
-				&genetic_analysis_order.customer_id,
-				&Self::account_id(),
-				genetic_analysis_order.total_price,
-				ExistenceRequirement::AllowDeath,
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		} else {
-			let asset_id = Self::asset_id(&genetic_analysis_order.currency)?;
-			match <T::Assets as fungibles::Mutate<T::AccountId>>::teleport(
-				asset_id,
-				&genetic_analysis_order.customer_id,
-				&Self::account_id(),
-				genetic_analysis_order.total_price.saturated_into(),
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		}
-
-		let genetic_analysis_order = Self::update_genetic_analysis_order_status(
-			genetic_analysis_order_id,
-			GeneticAnalysisOrderStatus::Paid,
-		);
-
-		Ok(genetic_analysis_order.unwrap())
-	}
-
-	fn fulfill_genetic_analysis_order(
-		escrow_account_id: &T::AccountId,
-		genetic_analysis_order_id: &T::Hash,
-	) -> Result<Self::GeneticAnalysisOrder, Self::Error> {
-		// Only the admin can fulfill the genetic_analysis_order
-		if escrow_account_id.clone() != EscrowKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		let genetic_analysis_order = GeneticAnalysisOrders::<T>::get(genetic_analysis_order_id);
-		if genetic_analysis_order.is_none() {
-			return Err(Error::<T>::GeneticAnalysisOrderNotFound)
-		}
-		let genetic_analysis_order = genetic_analysis_order.unwrap();
-
-		let genetic_analysis = T::GeneticAnalysis::genetic_analysis_by_genetic_analysis_tracking_id(
-			&genetic_analysis_order.genetic_analysis_tracking_id,
-		);
-		if !genetic_analysis.unwrap().process_success() {
-			return Err(Error::<T>::GeneticAnalysisNotSuccessfullyProcessed)
-		}
-
-		if !Self::is_pallet_balance_sufficient_for_transfer(genetic_analysis_order.total_price) {
-			return Err(Error::<T>::InsufficientPalletFunds)
-		}
-
-		// Calculate 5% of the price_component_value
-		let mut price_component_substracted_value: BalanceOf<T> = 0u128.saturated_into();
-		for analysis_order in genetic_analysis_order.prices.iter() {
-			price_component_substracted_value += analysis_order.value / 20u128.saturated_into();
-		}
-
-		// 5% of the price_component_value is substracted
-		let total_price_paid =
-			genetic_analysis_order.total_price - price_component_substracted_value;
-
-		// Withhold 5% for DBIO
-		if genetic_analysis_order.currency == CurrencyType::DBIO {
-			match CurrencyOf::<T>::transfer(
-				&Self::account_id(),
-				&genetic_analysis_order.seller_id,
-				total_price_paid,
-				ExistenceRequirement::KeepAlive,
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		} else {
-			let asset_id = Self::asset_id(&genetic_analysis_order.currency)?;
-			match <T::Assets as fungibles::Mutate<T::AccountId>>::teleport(
-				asset_id,
-				&Self::account_id(),
-				&genetic_analysis_order.seller_id,
-				total_price_paid.saturated_into(),
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		}
-
-		// Transfer 5% to DBIO Treasury
-		if genetic_analysis_order.currency == CurrencyType::DBIO {
-			match CurrencyOf::<T>::transfer(
-				&Self::account_id(),
-				&TreasuryKey::<T>::get().unwrap(),
-				price_component_substracted_value,
-				ExistenceRequirement::AllowDeath,
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		} else {
-			let asset_id = Self::asset_id(&genetic_analysis_order.currency)?;
-			match <T::Assets as fungibles::Mutate<T::AccountId>>::teleport(
-				asset_id,
-				&Self::account_id(),
-				&TreasuryKey::<T>::get().unwrap(),
-				price_component_substracted_value.saturated_into(),
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		}
-
-		let genetic_analysis_order = Self::update_genetic_analysis_order_status(
-			genetic_analysis_order_id,
-			GeneticAnalysisOrderStatus::Fulfilled,
-		);
-
-		Ok(genetic_analysis_order.unwrap())
-	}
-
-	fn set_genetic_analysis_order_refunded(
-		escrow_account_id: &T::AccountId,
-		genetic_analysis_order_id: &T::Hash,
-	) -> Result<Self::GeneticAnalysisOrder, Self::Error> {
-		if escrow_account_id.clone() != EscrowKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		let genetic_analysis_order = GeneticAnalysisOrders::<T>::get(genetic_analysis_order_id);
-		if genetic_analysis_order.is_none() {
-			return Err(Error::<T>::GeneticAnalysisOrderNotFound)
-		}
-
-		let genetic_analysis_order_can_be_refunded =
-			Self::genetic_analysis_order_can_be_refunded(genetic_analysis_order.clone().unwrap());
-		if !genetic_analysis_order_can_be_refunded {
-			return Err(Error::<T>::GeneticAnalysisOrderNotYetExpired)
-		}
-
-		let genetic_analysis_order = genetic_analysis_order.unwrap();
-		if !Self::is_pallet_balance_sufficient_for_transfer(genetic_analysis_order.total_price) {
-			return Err(Error::<T>::InsufficientPalletFunds)
-		}
-
-		// Transfer 5% to DBIO Treasury
-		if genetic_analysis_order.currency == CurrencyType::DBIO {
-			match CurrencyOf::<T>::transfer(
-				&Self::account_id(),
-				&genetic_analysis_order.customer_id,
-				genetic_analysis_order.total_price,
-				ExistenceRequirement::AllowDeath,
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		} else {
-			let asset_id = Self::asset_id(&genetic_analysis_order.currency)?;
-			match <T::Assets as fungibles::Mutate<T::AccountId>>::teleport(
-				asset_id,
-				&Self::account_id(),
-				&genetic_analysis_order.customer_id,
-				genetic_analysis_order.total_price.saturated_into(),
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		}
-
-		let genetic_analysis_order = Self::update_genetic_analysis_order_status(
-			genetic_analysis_order_id,
-			GeneticAnalysisOrderStatus::Refunded,
-		);
-		Ok(genetic_analysis_order.unwrap())
-	}
-
-	fn update_escrow_key(
-		account_id: &T::AccountId,
-		escrow_key: &T::AccountId,
-	) -> Result<(), Self::Error> {
-		if account_id.clone() != EscrowKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		EscrowKey::<T>::put(escrow_key);
-
-		Ok(())
-	}
-
-	fn update_treasury_key(
-		account_id: &T::AccountId,
-		treasury_key: &T::AccountId,
-	) -> Result<(), Self::Error> {
-		if account_id.clone() != TreasuryKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		TreasuryKey::<T>::put(treasury_key);
-
-		Ok(())
-	}
-
-	fn is_pending_genetic_analysis_order_ids_by_seller_exist(account_id: &T::AccountId) -> bool {
-		match PendingGeneticAnalysisOrdersBySeller::<T>::get(account_id) {
-			Some(_arr) => !_arr.is_empty(),
-			None => false,
-		}
-	}
-}
-
-impl<T: Config> Pallet<T> {
-	pub fn generate_genetic_analysis_order_id(
-		customer_id: &T::AccountId,
-		service_id: &T::Hash,
-	) -> T::Hash {
-		let mut customer_id_bytes = customer_id.encode();
-		let mut service_id_bytes = service_id.encode();
-		let account_info = frame_system::Pallet::<T>::account(customer_id);
-		let mut nonce_bytes = account_info.nonce.encode();
-
-		customer_id_bytes.append(&mut service_id_bytes);
-		customer_id_bytes.append(&mut nonce_bytes);
-
-		let seed = &customer_id_bytes;
-		T::Hashing::hash(seed)
-	}
-
-	pub fn update_genetic_analysis_order_status(
-		genetic_analysis_order_id: &T::Hash,
-		status: GeneticAnalysisOrderStatus,
-	) -> Option<GeneticAnalysisOrderOf<T>> {
-		GeneticAnalysisOrders::<T>::mutate(genetic_analysis_order_id, |genetic_analysis_order| {
-			match genetic_analysis_order {
-				None => None,
-				Some(genetic_analysis_order) => {
-					genetic_analysis_order.status = status;
-					genetic_analysis_order.updated_at = pallet_timestamp::Pallet::<T>::get();
-					Some(genetic_analysis_order.clone())
-				},
-			}
-		})
-	}
-
-	pub fn insert_genetic_analysis_order_to_storage(
-		genetic_analysis_order: &GeneticAnalysisOrderOf<T>,
-	) {
-		GeneticAnalysisOrders::<T>::insert(genetic_analysis_order.id, genetic_analysis_order);
-		LastGeneticAnalysisOrderByCustomer::<T>::insert(
-			&genetic_analysis_order.customer_id,
-			genetic_analysis_order.id,
-		);
-		Self::insert_genetic_analysis_order_id_into_genetic_analysis_orders_by_seller(
-			genetic_analysis_order,
-		);
-		Self::insert_genetic_analysis_order_id_into_pending_genetic_analysis_orders_by_seller(
-			genetic_analysis_order,
-		);
-		Self::insert_genetic_analysis_order_id_into_genetic_analysis_orders_by_customer(
-			genetic_analysis_order,
-		);
-	}
-
-	pub fn insert_genetic_analysis_order_id_into_genetic_analysis_orders_by_seller(
-		genetic_analysis_order: &GeneticAnalysisOrderOf<T>,
-	) {
-		match GeneticAnalysisOrdersBySeller::<T>::get(&genetic_analysis_order.seller_id) {
-			None => {
-				GeneticAnalysisOrdersBySeller::<T>::insert(
-					&genetic_analysis_order.seller_id,
-					vec![genetic_analysis_order.id],
-				);
-			},
-			Some(mut genetic_analysis_orders) => {
-				genetic_analysis_orders.push(genetic_analysis_order.id);
-				GeneticAnalysisOrdersBySeller::<T>::insert(
-					&genetic_analysis_order.seller_id,
-					genetic_analysis_orders,
-				);
-			},
-		}
-	}
-
-	pub fn insert_genetic_analysis_order_id_into_genetic_analysis_orders_by_customer(
-		genetic_analysis_order: &GeneticAnalysisOrderOf<T>,
-	) {
-		match GeneticAnalysisOrdersByCustomer::<T>::get(&genetic_analysis_order.customer_id) {
-			None => {
-				GeneticAnalysisOrdersByCustomer::<T>::insert(
-					&genetic_analysis_order.customer_id,
-					vec![genetic_analysis_order.id],
-				);
-			},
-			Some(mut genetic_analysis_orders) => {
-				genetic_analysis_orders.push(genetic_analysis_order.id);
-				GeneticAnalysisOrdersByCustomer::<T>::insert(
-					&genetic_analysis_order.customer_id,
-					genetic_analysis_orders,
-				);
-			},
-		}
-	}
-
-	pub fn insert_genetic_analysis_order_id_into_pending_genetic_analysis_orders_by_seller(
-		genetic_analysis_order: &GeneticAnalysisOrderOf<T>,
-	) {
-		match PendingGeneticAnalysisOrdersBySeller::<T>::get(&genetic_analysis_order.seller_id) {
-			None => {
-				PendingGeneticAnalysisOrdersBySeller::<T>::insert(
-					&genetic_analysis_order.seller_id,
-					vec![genetic_analysis_order.id],
-				);
-			},
-			Some(mut genetic_analysis_orders) => {
-				genetic_analysis_orders.push(genetic_analysis_order.id);
-				PendingGeneticAnalysisOrdersBySeller::<T>::insert(
-					&genetic_analysis_order.seller_id,
-					genetic_analysis_orders,
-				);
-			},
-		}
-	}
-
-	pub fn remove_genetic_analysis_order_id_from_pending_genetic_analysis_orders_by_seller(
-		seller_id: &T::AccountId,
-		genetic_analysis_order_id: &T::Hash,
-	) {
-		let mut genetic_analysis_orders =
-			PendingGeneticAnalysisOrdersBySeller::<T>::get(seller_id).unwrap_or_default();
-		genetic_analysis_orders.retain(|o_id| o_id != genetic_analysis_order_id);
-		PendingGeneticAnalysisOrdersBySeller::<T>::insert(seller_id, genetic_analysis_orders);
-	}
-
-	pub fn genetic_analysis_order_can_be_refunded(
-		genetic_analysis_order: GeneticAnalysisOrderOf<T>,
-	) -> bool {
-		let genetic_analysis =
-			T::GeneticAnalysis::genetic_analysis_by_genetic_analysis_tracking_id(
-				&genetic_analysis_order.genetic_analysis_tracking_id,
-			)
-			.unwrap();
-		if !genetic_analysis.is_rejected() {
-			return false
-		}
-		true
-	}
-
-	/// The injected pallet ID
-	pub fn get_pallet_id() -> AccountIdOf<T> {
-		T::PalletId::get().into_account()
-	}
-
-	/// The account ID that holds the funds
-	pub fn account_id() -> AccountIdOf<T> {
-		<PalletAccount<T>>::get().unwrap()
-	}
-
-	/// Is the balance sufficient for payment
-	pub fn is_balance_sufficient_for_payment(
-		account_id: &AccountIdOf<T>,
-		price: BalanceOf<T>,
-	) -> bool {
-		let balance = T::Currency::free_balance(account_id);
-		balance >= price
-	}
-
-	/// Is the pallet balance sufficient for transfer
-	pub fn is_pallet_balance_sufficient_for_transfer(price: BalanceOf<T>) -> bool {
-		let balance = T::Currency::free_balance(&Self::account_id());
-		balance >= price
-	}
-
-	/// Set current escrow amount
-	pub fn set_escrow_amount() {
-		TotalEscrowAmount::<T>::put(T::Currency::free_balance(&Self::account_id()));
-	}
-
-	// Get token identifier
-	pub fn asset_id(currency_type: &CurrencyType) -> Result<u32, Error<T>> {
-		currency_type
-			.to_asset_id()
-			.parse::<u32>()
-			.map_err(|_| Error::<T>::WrongAssetIdFormat)
-	}
-}
-
-impl<T: Config> GeneticAnalysisOrderEventEmitter<T> for Pallet<T> {
-	fn emit_event_genetic_analysis_order_failed(genetic_analysis_order_id: &HashOf<T>) {
-		match Self::genetic_analysis_order_by_id(genetic_analysis_order_id) {
-			None => Self::deposit_event(Event::GeneticAnalysisOrderNotFound),
-			Some(genetic_analysis_order) =>
-				Self::deposit_event(Event::GeneticAnalysisOrderFailed(genetic_analysis_order)),
-		}
-	}
-}
-
-impl<T: Config> GeneticAnalysisOrderStatusUpdater<T> for Pallet<T> {
-	fn update_status_failed(genetic_analysis_order_id: &HashOf<T>) {
-		match Self::genetic_analysis_order_by_id(genetic_analysis_order_id) {
-			None => Self::deposit_event(Event::GeneticAnalysisOrderNotFound),
-			Some(genetic_analysis_order) => {
-				Self::update_genetic_analysis_order_status(
-					&genetic_analysis_order.id,
-					GeneticAnalysisOrderStatus::Failed,
-				);
-			},
-		}
-	}
-
-	fn remove_genetic_analysis_order_id_from_pending_genetic_analysis_orders_by_seller(
-		seller_id: &AccountIdOf<T>,
-		genetic_analysis_order_id: &HashOf<T>,
-	) {
-		Self::remove_genetic_analysis_order_id_from_pending_genetic_analysis_orders_by_seller(
-			seller_id,
-			genetic_analysis_order_id,
-		);
-	}
-
-	fn is_pending_genetic_analysis_order_by_seller_exist(seller_id: &AccountIdOf<T>) -> bool {
-		Self::is_pending_genetic_analysis_order_ids_by_seller_exist(seller_id)
 	}
 }
