+++ pallets/genetic-analysis/tests/src/lib.rs
@@ -82,6 +82,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -185,6 +186,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -264,6 +266,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -379,6 +382,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -480,6 +484,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -586,6 +591,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -664,6 +670,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
@@ -761,6 +768,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			assert_noop!(
@@ -837,6 +845,7 @@ mod tests {
 				0,
 				Keccak256::hash("0xhJ7TRe456FADD2726A132ABJK5RCc9E6fC5869F4".as_bytes()),
 				"DeBio Genetic Genetic Link".as_bytes().to_vec(),
+				None,
 			));
 
 			let _genetic_analysis =
